MyselfTestを分離
