+++ Assets/Editor/TEST.MyselfTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1275f0f804b2dbf4ba548bfed48023f2
+guid: d7333f0482c53744a944abb8ab77e9d7
 timeCreated: 1513340559
 licenseType: Free
 MonoImporter:
